move stories :arrow_right: examples for :racehorse:
and not include in any other bundles
