+++ .travis.yml
@@ -19,8 +19,10 @@ script: npm run ci
 cache:
   directories:
     - $HOME/.electron
+
 after_success:
   - npm run storybook:ghpages
+  
 deploy:
   provider: pages
   skip-cleanup: true
